+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ unreleased
 ----------
 * Add the ``da.pr.set()`` and ``ds.pr.set()`` functions for overwriting / infilling /
   extending data.
+* Allow for more than one source in a single Dataset / DataArray.
 
 0.3.1
 -----